requirements.txt
flask==3.1.3
flask-cors==6.0.2
supabase==2.7.4
gotrue==2.7.0
postgrest==0.16.11
websockets>=14
python-dotenv==1.2.2
PyJWT==2.12.1
stripe==15.0.0
gunicorn==25.3.0
